refactor: ReservationController의 어드민 API 분리
