+++ src/main/java/roomescape/controller/reservation/ReservationController.java
@@ -1,16 +1,12 @@
-package roomescape.controller;
+package roomescape.controller.reservation;
 
 import jakarta.validation.Valid;
 import java.net.URI;
-import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.service.ReservationService;
 import roomescape.service.dto.AuthInfo;
@@ -27,11 +23,6 @@ public class ReservationController {
         this.reservationService = reservationService;
     }
 
-    @GetMapping("/reservations")
-    public ResponseEntity<List<ReservationResponse>> findAllReservations() {
-        return ResponseEntity.ok(reservationService.findAll());
-    }
-
     @GetMapping("/reservations-mine")
     public ResponseEntity<List<MyReservationResponse>> findMyReservations(AuthInfo authInfo) {
         return ResponseEntity.ok(reservationService.findMyReservations(authInfo));
@@ -46,28 +37,5 @@ public class ReservationController {
                 authInfo.id());
         ReservationResponse reservationResponse = reservationService.save(reservationCreateRequest);
         return ResponseEntity.created(URI.create("admin/reservation")).body(reservationResponse);
-    }
-
-    @PostMapping("/admin/reservations")
-    public ResponseEntity<ReservationResponse> createReservation(
-            @Valid @RequestBody ReservationCreateRequest reservationCreateRequest) {
-        ReservationResponse reservationResponse = reservationService.save(reservationCreateRequest);
-        return ResponseEntity.created(URI.create("admin/reservation")).body(reservationResponse);
-    }
-
-    @GetMapping("/admin/reservations")
-    public ResponseEntity<List<ReservationResponse>> findBy(
-            @RequestParam(required = false, value = "themeId") Long themeId,
-            @RequestParam(required = false, value = "memberId") Long memberId,
-            @RequestParam(required = false, value = "dateFrom") LocalDate dateFrom,
-            @RequestParam(required = false, value = "dateTo") LocalDate dateTo
-    ) {
-        return ResponseEntity.ok().body(reservationService.findBy(themeId, memberId, dateFrom, dateTo));
-    }
-
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        reservationService.deleteById(id);
-        return ResponseEntity.noContent().build();
     }
 }
